Split arbenv/gpu_api.hpp
